requirements.txt
lxml_html_clean
ollama
psutil
requests-html
scapy
